refactorig step_3, update designe step_3
